third_party/WebKit/LayoutTests/fast/dom/shadow/css-focus-pseudo-match-shadow-host4-expected.txt

CONSOLE WARNING: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated. See https://www.chromestatus.com/features/4668884095336448 for more details.
Testing shadow host :focus and distributed nodes
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
(1/3) shadow host without tabindex
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
(2/3) shadow host with tabindex=-1
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
(3/3) shadow host with tabindex=0
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
